Fixed path for hohmann_transfer.py + pylint error
